disable dfs sensor backup for secure environment

+++ hdi/hdinsight-unravel-spark-script-action/unravel_hdi_spark_bootstrap_4.5.sh
@@ -912,24 +912,27 @@ function es_install() {
   wget -4 -q -T 10 -t 5 -O - $UESURL > ${TMP_DIR}/$UES_JAR_NAME
   RC=$?
 
-  if [ $RC -ne 0 ]; then
-    download_from_dfs $UES_JAR_NAME ${TMP_DIR}/$UES_JAR_NAME
-    RC=$?
-  fi
-
-  if [ $RC -eq 0 ]; then
-      upload_to_dfs ${TMP_DIR}/$UES_JAR_NAME
-      sudo /bin/cp ${TMP_DIR}/$UES_JAR_NAME  ${UES_PATH}
-      [ -d "${UES_PATH}/dlib" ] && rm -rf ${UES_PATH}/dlib
-      sudo unzip -o /usr/local/unravel_es/$UES_JAR_NAME -d ${UES_PATH}/
-      sudo chmod 755 ${UES_PATH}/dbin/*
-      sudo chown -R "${UNRAVEL_ES_USER}":"${UNRAVEL_ES_GROUP}" ${UES_PATH}
-  else
-      echo "ERROR: Fetch of $UESURL failed, RC=$RC" |tee -a $OUT_FILE
-      download_from_dfs $UES_JAR_NAME
-      if [ $? -ne 0 ]; then
-        return 1
-      fi
+  is_secure
+  if [[ $? -eq 1 ]];then
+    if [ $RC -ne 0 ]; then
+      download_from_dfs $UES_JAR_NAME ${TMP_DIR}/$UES_JAR_NAME
+      RC=$?
+    fi
+  
+    if [ $RC -eq 0 ]; then
+        upload_to_dfs ${TMP_DIR}/$UES_JAR_NAME
+        sudo /bin/cp ${TMP_DIR}/$UES_JAR_NAME  ${UES_PATH}
+        [ -d "${UES_PATH}/dlib" ] && rm -rf ${UES_PATH}/dlib
+        sudo unzip -o /usr/local/unravel_es/$UES_JAR_NAME -d ${UES_PATH}/
+        sudo chmod 755 ${UES_PATH}/dbin/*
+        sudo chown -R "${UNRAVEL_ES_USER}":"${UNRAVEL_ES_GROUP}" ${UES_PATH}
+    else
+        echo "ERROR: Fetch of $UESURL failed, RC=$RC" |tee -a $OUT_FILE
+        download_from_dfs $UES_JAR_NAME
+        if [ $? -ne 0 ]; then
+          return 1
+        fi
+    fi
   fi
 
   # start
